move manager offsets to new files
